+++ src/main/webapp/resources/parsleyjs/src/parsley/factory.js
@@ -56,7 +56,6 @@ ParsleyFactory.prototype = {
   // Multiples fields are a real nightmare :(
   // Maybe some refactoring would be appreciated here...
   handleMultiple: function () {
-    var that = this;
     var name;
     var multiple;
     var parsleyMultipleInstance;
@@ -85,14 +84,14 @@ ParsleyFactory.prototype = {
 
     // Add proper `data-parsley-multiple` to siblings if we have a valid multiple name
     if ('undefined' !== typeof name) {
-      $('input[name="' + name + '"]').each(function () {
-        if ($(this).is('input[type=radio], input[type=checkbox]'))
-          $(this).attr(that.options.namespace + 'multiple', that.options.multiple);
+      $('input[name="' + name + '"]').each((i, input) => {
+        if ($(input).is('input[type=radio], input[type=checkbox]'))
+          $(input).attr(this.options.namespace + 'multiple', this.options.multiple);
       });
     }
 
     // Check here if we don't already have a related multiple instance saved
-    var $previouslyRelated = this._findRelatedMultiple();
+    var $previouslyRelated = this._findRelated();
     for (var i = 0; i < $previouslyRelated.length; i++) {
       parsleyMultipleInstance = $($previouslyRelated.get(i)).data('Parsley');
       if ('undefined' !== typeof parsleyMultipleInstance) {