+++ src/main/webapp/resources/parsleyjs/src/parsley/factory/constraint.js
@@ -4,7 +4,7 @@ import ParsleyValidator from '../validator';
 
 
 var ConstraintFactory = function (parsleyField, name, requirements, priority, isDomConstraint) {
-  if (!new RegExp('ParsleyField').test(parsleyField.__class__))
+  if (!/ParsleyField/.test(parsleyField.__class__))
     throw new Error('ParsleyField or ParsleyFieldMultiple instance expected');
 
   var validatorSpec = window.Parsley._validatorRegistry.validators[name];
@@ -34,9 +34,8 @@ ConstraintFactory.prototype = {
   },
 
   _parseRequirements: function(options) {
-    var that = this;
-    this.requirementList = this.validator.parseRequirements(this.requirements, function(key) {
-      return options[that.name + capitalize(key)];
+    this.requirementList = this.validator.parseRequirements(this.requirements, key => {
+      return options[this.name + capitalize(key)];
     });
   }
 };