+++ src/main/webapp/resources/parsleyjs/src/parsley/form.js
@@ -19,8 +19,6 @@ var statusMapping = {pending: null, resolved: true, rejected: false};
 
 ParsleyForm.prototype = {
   onSubmitValidate: function (event) {
-    var that = this;
-
     // This is a Parsley generated submit event, do not validate, do not prevent, simply exit and keep normal behavior
     if (true === event.parsley)
       return;
@@ -38,9 +36,9 @@ ParsleyForm.prototype = {
     event.stopImmediatePropagation();
     event.preventDefault();
 
-    this.whenValidate(undefined, undefined, event)
-      .done(function () { that._submit(); })
-      .always(function () { that._$submitSource = null; });
+    this.whenValidate({event})
+      .done(() => { this._submit(); })
+      .always(() => { this._$submitSource = null; });
 
     return this;
   },
@@ -67,17 +65,27 @@ ParsleyForm.prototype = {
   },
 
   // Performs validation on fields while triggering events.
-  // @returns `true` if al validations succeeds, `false`
+  // @returns `true` if all validations succeeds, `false`
   // if a failure is immediately detected, or `null`
   // if dependant on a promise.
-  // Prefer `whenValidate`.
-  validate: function (group, force, event) {
-    return statusMapping[ this.whenValidate(group, force, event).state() ];
+  // Consider using `whenValidate` instead.
+  validate: function (options) {
+    if (arguments.length >= 1 && !$.isPlainObject(options)) {
+      ParsleyUtils.warnOnce('Calling validate on a parsley form without passing arguments as an object is deprecated.');
+      var [group, force, event] = arguments;
+      options = {group, force, event};
+    }
+    return statusMapping[ this.whenValidate(options).state() ];
   },
 
-  whenValidate: function (group, force, event) {
-    var that = this;
+  whenValidate: function ({group, force, event} = {}) {
     this.submitEvent = event;
+    if (event) {
+      this.submitEvent.preventDefault = () => {
+        ParsleyUtils.warnOnce("Using `this.submitEvent.preventDefault()` is deprecated; instead, call `this.validationResult = false`");
+        this.validationResult = false;
+      };
+    }
     this.validationResult = true;
 
     // fire validate event to eventually modify things before very validation
@@ -86,25 +94,23 @@ ParsleyForm.prototype = {
     // Refresh form DOM options and form's fields that could have changed
     this._refreshFields();
 
-    var promises = this._withoutReactualizingFormOptions(function () {
-      return $.map(this.fields, function(field) {
-        // do not validate a field if not the same as given validation group
-        if (!group || that._isFieldInGroup(field, group))
-          return field.whenValidate(force);
+    var promises = this._withoutReactualizingFormOptions(() => {
+      return $.map(this.fields, field => {
+        return field.whenValidate({force, group});
       });
     });
 
-    var promiseBasedOnValidationResult = function () {
+    var promiseBasedOnValidationResult = () => {
       var r = $.Deferred();
-      if (false === that.validationResult)
+      if (false === this.validationResult)
         r.reject();
       return r.resolve().promise();
     };
 
-    return $.when.apply($, promises)
-      .done(  function () { that._trigger('success'); })
-      .fail(  function () { that.validationResult = false; that._trigger('error'); })
-      .always(function () { that._trigger('validated'); })
+    return $.when(...promises)
+      .done(  () => { this._trigger('success'); })
+      .fail(  () => { this.validationResult = false; this._trigger('error'); })
+      .always(() => { this._trigger('validated'); })
       .pipe(  promiseBasedOnValidationResult, promiseBasedOnValidationResult);
   },
 
@@ -112,31 +118,27 @@ ParsleyForm.prototype = {
   // Returns `true` if all fields are valid, `false` if a failure is detected
   // or `null` if the result depends on an unresolved promise.
   // Prefer using `whenValid` instead.
-  isValid: function (group, force) {
-    return statusMapping[ this.whenValid(group, force).state() ];
+  isValid: function (options) {
+    if (arguments.length >= 1 && !$.isPlainObject(options)) {
+      ParsleyUtils.warnOnce('Calling isValid on a parsley form without passing arguments as an object is deprecated.');
+      var [group, force] = arguments;
+      options = {group, force};
+    }
+    return statusMapping[ this.whenValid(options).state() ];
   },
 
   // Iterate over refreshed fields and validate them.
   // Returns a promise.
   // A validation that immediately fails will interrupt the validations.
-  whenValid: function (group, force) {
-    var that = this;
+  whenValid: function ({group, force} = {}) {
     this._refreshFields();
 
-    var promises = this._withoutReactualizingFormOptions(function () {
-      return $.map(this.fields, function(field) {
-        // do not validate a field if not the same as given validation group
-        if (!group || that._isFieldInGroup(field, group))
-          return field.whenValid(force);
+    var promises = this._withoutReactualizingFormOptions(() => {
+      return $.map(this.fields, field => {
+        return field.whenValid({group, force});
       });
     });
-    return $.when.apply($, promises);
-  },
-
-  _isFieldInGroup: function (field, group) {
-    if ($.isArray(field.options.group))
-      return -1 !== $.inArray(group, field.options.group);
-    return field.options.group === group;
+    return $.when(...promises);
   },
 
   _refreshFields: function () {
@@ -144,29 +146,28 @@ ParsleyForm.prototype = {
   },
 
   _bindFields: function () {
-    var self = this;
     var oldFields = this.fields;
 
     this.fields = [];
     this.fieldsMappedById = {};
 
-    this._withoutReactualizingFormOptions(function () {
+    this._withoutReactualizingFormOptions(() => {
       this.$element
       .find(this.options.inputs)
       .not(this.options.excluded)
-      .each(function () {
-        var fieldInstance = new window.Parsley.Factory(this, {}, self);
+      .each((_, element) => {
+        var fieldInstance = new window.Parsley.Factory(element, {}, this);
 
         // Only add valid and not excluded `ParsleyField` and `ParsleyFieldMultiple` children
         if (('ParsleyField' === fieldInstance.__class__ || 'ParsleyFieldMultiple' === fieldInstance.__class__) && (true !== fieldInstance.options.excluded))
-          if ('undefined' === typeof self.fieldsMappedById[fieldInstance.__class__ + '-' + fieldInstance.__id__]) {
-            self.fieldsMappedById[fieldInstance.__class__ + '-' + fieldInstance.__id__] = fieldInstance;
-            self.fields.push(fieldInstance);
+          if ('undefined' === typeof this.fieldsMappedById[fieldInstance.__class__ + '-' + fieldInstance.__id__]) {
+            this.fieldsMappedById[fieldInstance.__class__ + '-' + fieldInstance.__id__] = fieldInstance;
+            this.fields.push(fieldInstance);
           }
       });
 
-      $(oldFields).not(self.fields).each(function () {
-        this._trigger('reset');
+      $(oldFields).not(this.fields).each((_, field) => {
+        field._trigger('reset');
       });
     });
     return this;
@@ -182,7 +183,7 @@ ParsleyForm.prototype = {
   _withoutReactualizingFormOptions: function (fn) {
     var oldActualizeOptions = this.actualizeOptions;
     this.actualizeOptions = function () { return this; };
-    var result = fn.call(this); // Keep the current `this`.
+    var result = fn();
     this.actualizeOptions = oldActualizeOptions;
     return result;
   },