+++ src/main/webapp/resources/parsleyjs/src/parsley/multiple.js
@@ -46,18 +46,20 @@ ParsleyMultiple.prototype = {
   // See `ParsleyField.getValue()`
   getValue: function () {
     // Value could be overriden in DOM
-    if ('undefined' !== typeof this.options.value)
+    if ('function' === typeof this.options.value)
+      value = this.options.value(this);
+    else if ('undefined' !== typeof this.options.value)
       return this.options.value;
 
     // Radio input case
     if (this.$element.is('input[type=radio]'))
-      return this._findRelatedMultiple().filter(':checked').val() || '';
+      return this._findRelated().filter(':checked').val() || '';
 
     // checkbox input case
     if (this.$element.is('input[type=checkbox]')) {
       var values = [];
 
-      this._findRelatedMultiple().filter(':checked').each(function () {
+      this._findRelated().filter(':checked').each(function () {
         values.push($(this).val());
       });